refactor: モノリシックコンポーネントからhookを抽出 (Phase 1-2)
5326行のcomm-time.tsxから3つのカスタムhookとユーティリティを抽出:
- useAlarmSystem: アラーム/通知/振動/フラッシュ/チクタク音 (436行)
- useMeetingTimer: ミーティングタイマー状態機械 (365行)
- usePomodoroTimer: ポモドーロタイマー状態機械 (239行)
- types/index.ts: 共有型定義を集約
- lib/constants.ts: 初期値定数を分離

comm-time.tsx: 5326行 → 4450行 (-876行)
見た目・機能の変更なし
