+++ .gitignore
@@ -37,3 +37,4 @@ next-env.d.ts
 
 # claude
 .claude/settings.local.json
+.claude/skills/ui-ux-pro-max/
